Add Dockerfile support:
We now have a docker file. It includes instructions on building and
running. The `README.md` is updated on how to use it.

It also includes the ability to build it without the clients (since
you'll probably run the clients locally).

Without clients, I got it down to 136MB on my laptop.